Update canvas creation and square resolution
The resetButton event listener now prompts the user to input the desired
number of squares per side, with a maximum of 100 squares. The
createCanvas function checks if the user input exceeds the maximum, and
if so, sets the number of squares to 100. the getResolution function
sets the width and height of each square to be proportional to the
number of squares per side. THe squareDiv.setAttribute function call in createCanvas was replaced with a call to the new getResolution function.